Fix code review issues from PR #100
Critical fixes:
- Guard onCanvasClicked against nullptr canvas when VideoSurface active
- Reset video surfaces before lv_obj_clean to prevent double-delete
- Move static locals in UI SimRunning to struct members to avoid stale state

Bug fixes:
- Unify APU sweep logic with proper shift/bounds guards
- Extract NesApuGet::Okay::fromSnapshot() to eliminate duplicated mapping
- Remove dead s16Buffer_ from NesAudioPlayer
- Clear bestVideoFrame_ on scenario switch
- Replace static nextId atomics with allocateRequestId()
- Remove redundant cleanup() before reset() on VideoSurface
- Fix VideoSurface::getCanvasHeight return type to uint32_t
- Reuse presentStartMs instead of second monotonicNowMs() call
- Switch MessageParser from raw spdlog to LOG_* macros
- Add const to local variables, fix typo in onExit comment
- Alphabetize TrainingResult entries in ApiCommand.h and Event.h

Cleanup:
- Extract LoopTimingStats for display backend run loops
- Remove duplicate extern, unused includes, stdbool.h from C++ files
- Remove empty section headers and @param doc comments in backends
- Move SMOLNES_APU_SAMPLE_COPY_MAX to SmolnesRuntimeBackend.c
- Remove redundant SimPlayground forward declaration

+++ apps/src/ui/rendering/VideoSurface.cpp
@@ -86,7 +86,7 @@ uint32_t VideoSurface::getCanvasWidth() const
     return canvasWidth_;
 }
 
-uint16_t VideoSurface::getCanvasHeight() const
+uint32_t VideoSurface::getCanvasHeight() const
 {
     return canvasHeight_;
 }
